requirements.txt
Click==7.0
colander==1.7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
iso8601==0.1.8
itsdangerous==1.1.0
Jinja2==2.10.3
lark-parser==0.7.4
MarkupSafe==1.1.1
pep8==1.7.1
pycodestyle==2.5.0
redis==2.8.0
SQLAlchemy==1.3.7
translationstring==1.3
Werkzeug==0.16.0
WTForms==2.2.1
uwsgi